test: extract manage tasks route fixtures
